comments working properly. must format date

+++ services/src/main/java/ar/edu/itba/paw/services/CommentServiceImpl.java
@@ -5,6 +5,7 @@ import ar.edu.itba.paw.persistence.CommentDao;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import ar.edu.itba.paw.persistence.ReviewDao;
 
 import java.util.List;
 import java.util.Optional;
@@ -13,10 +14,12 @@ import java.util.Optional;
 public class CommentServiceImpl implements CommentService {
 
     private final CommentDao commentDao;
+    private final ReviewDao reviewDao;
 
     @Autowired
-    public CommentServiceImpl(final CommentDao commentDao) {
+    public CommentServiceImpl(final CommentDao commentDao, final ReviewDao reviewDao) {
         this.commentDao = commentDao;
+        this.reviewDao = reviewDao;
     }
 
     @Override
@@ -32,12 +35,24 @@ public class CommentServiceImpl implements CommentService {
     @Override
     @Transactional
     public Comment save(Comment comment) {
-        return commentDao.save(comment);
+        Comment savedComment = commentDao.save(comment);
+        updateReviewCommentAmount(comment.getReviewId());
+        return savedComment;
     }
 
     @Override
     @Transactional
     public void deleteById(long id) {
-        commentDao.deleteById(id);
+        Optional<Comment> comment = commentDao.findById(id);
+        if (comment.isPresent()) {
+            commentDao.deleteById(id);
+            updateReviewCommentAmount(comment.get().getReviewId());
+        }
+    }
+
+    @Override
+    @Transactional
+    public void updateReviewCommentAmount(long reviewId) {
+        reviewDao.updateCommentAmount(reviewId);
     }
 }
